Coalesce G11: lift stable-KBO table and scattered-disk sim into p9-core
