+++ crates/p9-core/src/initial_conditions/mod.rs
@@ -1,2 +1,3 @@
 pub mod planets;
 pub mod scattered_disk;
+pub mod scattered_disk_sim;
